# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highleads API",
        "description": "API highleads-backend",
        "contact": {
            "name": "Marco Aurélio",
            "email": "marcoascabral@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/contatos-satus/": {
            "get": {
                "tags": [
                    "Contato"
                ],
                "summary": "Exibe os Contatos cadastrados filtrado por status",
                "description": "Lista os contatos cadastrados filtrados por status",
                "operationId": "a359bdeca55da6de4078da7414c14203",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "status",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados."
                    },
                    "404": {
                        "description": "Item não encontrado."
                    }
                }
            }
        },
        "/v1/contatos/{id}": {
            "get": {
                "tags": [
                    "Contato"
                ],
                "summary": "Exibe um contato",
                "description": "Exibe o contato por seu ID.",
                "operationId": "55c4e1670588f2f3392f115c33881a7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna dados do contato."
                    },
                    "404": {
                        "description": "Contato não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Contato"
                ],
                "summary": "Alterar um Contato",
                "description": "Alterar um contato por ID",
                "operationId": "f23f02ea4f9af93a64c3172c6cdeece1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Dados do Contato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "site",
                                    "nome",
                                    "email",
                                    "telefone",
                                    "remoteip",
                                    "datahora"
                                ],
                                "properties": {
                                    "site": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "https://zoit.com.br"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "José da Silva"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2199999999"
                                    },
                                    "remoteip": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "192.168.1.1"
                                    },
                                    "datahora": {
                                        "type": "datetime",
                                        "example": "2022-12-19 09:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contato salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir um Contato."
                    },
                    "404": {
                        "description": "Contato não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contato"
                ],
                "summary": "Exclui um contato",
                "description": "Exclui o contato por seu ID.",
                "operationId": "c299dcf95b8a240db8c6ed22848737f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do contato",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Contato excluído com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "excluído": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contato não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/contatos": {
            "post": {
                "tags": [
                    "Contato"
                ],
                "summary": "Criar um Contato",
                "description": "Criar um novo Contato",
                "operationId": "22caebf5fbc9694c0164b2430ba81e1f",
                "requestBody": {
                    "description": "Dados do Contato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "site",
                                    "nome",
                                    "email",
                                    "telefone",
                                    "remoteip",
                                    "datahora"
                                ],
                                "properties": {
                                    "site": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "https://zoit.com.br"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "José da Silva"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2199999999"
                                    },
                                    "remoteip": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "192.168.1.1"
                                    },
                                    "datahora": {
                                        "type": "datetime",
                                        "example": "2022-12-19 09:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contato salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir um Contato."
                    }
                }
            }
        },
...
    }
}